requirements.txt
python-whois
requests
colorama
dnspython
